# [FSDP] use all_gather for 10X OSD consolidation speedup