TLDR: Using `all_gather` instead of `broadcast` for optimizer state consolidation appears to be a speed win without a memory cost.

### Approach:
- for tensor state in OSD we use `all_gather`
- for non-tensor metadata (`loss_scale`, `param_groups`, `num_padded`) we use `broadcast`.
- The same OSD (for 300M param model) takes 110 ms to consolidate vs 2300 ms, speedup usually around 10X.
- assume `recipient_rank=0`, since there are no other callers. 


### Evidence of Win:
- This appears not to use extra GPU RAM, and save a lot of time. For the unittests, the process goes from 2300 MS to consolidate a larger MOE to 110 Ms. 
- For a large fairseq 2.2B param model (1.2T config), it takes 80s instead of 800s, with less CUDA usage.
- fairseq resumption test (resume training from ckpt should result as same loss as training from scratch) passes.